+++ tools/auth/auth.go
@@ -112,6 +112,8 @@ func NewProviderByName(name string) (Provider, error) {
 		return NewGiteeProvider(), nil
 	case NameLivechat:
 		return NewLivechatProvider(), nil
+	case NameAuthentik:
+		return NewAuthentikProvider(), nil
 	default:
 		return nil, errors.New("Missing provider " + name)
 	}