+++ tools/auth/auth_test.go
@@ -135,4 +135,13 @@ func TestNewProviderByName(t *testing.T) {
 	if _, ok := p.(*auth.Livechat); !ok {
 		t.Error("Expected to be instance of *auth.Livechat")
 	}
+
+	// authentik
+	p, err = auth.NewProviderByName(auth.NameAuthentik)
+	if err != nil {
+		t.Errorf("Expected nil, got error %v", err)
+	}
+	if _, ok := p.(*auth.Authentik); !ok {
+		t.Error("Expected to be instance of *auth.Authentik")
+	}
 }